feat: create image 정리
